Refactored testbase and Issue86Test In order to improve testing experience I moved major parts to testbase to make the UI tests itself more generic. Improved readability.
