+++ src/test/java/de/bayern/gdi/TestBase.java
@@ -22,15 +22,16 @@ import de.bayern.gdi.gui.Controller;
 import de.bayern.gdi.gui.DataBean;
 import de.bayern.gdi.gui.Start;
 import de.bayern.gdi.gui.WarningPopup;
-import de.bayern.gdi.utils.I18n;
 import de.bayern.gdi.utils.Config;
-import de.bayern.gdi.utils.Unauthorized;
-import de.bayern.gdi.utils.FileResponseHandler;
 import de.bayern.gdi.utils.DocumentResponseHandler;
+import de.bayern.gdi.utils.FileResponseHandler;
+import de.bayern.gdi.utils.I18n;
 import de.bayern.gdi.utils.Misc;
+import de.bayern.gdi.utils.Unauthorized;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.TextField;
 import javafx.scene.control.TitledPane;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
@@ -76,12 +77,25 @@ public class TestBase extends ApplicationTest {
      * Name of the Logo.
      */
     private static final String LOGONAME = "icon_118x118_300dpi.jpg";
-
+    /**
+     * Titlepane
+     */
+    private static final String HISTORY_PARENT =
+            "#logHistoryParent";
+    /**
+     * ServiceURL element
+     */
+    private static final String SERVICE_URL =
+            "#serviceURL";
+    /**
+     * Ready state
+     */
+    private static final String READY_STATUS =
+            "status.ready";
     /**
      * The Scene.
      */
     private Scene scene;
-
     /**
      * The logger.
      */
@@ -131,10 +145,11 @@ public class TestBase extends ApplicationTest {
 
     /**
      * Waits for Display of Status in TitlePane.
+     *
      * @param titlePane Pane to watch
-     * @param status Status to watch for
+     * @param status    Status to watch for
      */
-    void waitForStatus(TitledPane titlePane, String status) {
+    private void waitForStatus(TitledPane titlePane, String status) {
         await()
                 .atMost(TIMEOUT_SECONDS, SECONDS)
                 .pollInterval(POLL_MILLISECONDS, MILLISECONDS)
@@ -145,13 +160,58 @@ public class TestBase extends ApplicationTest {
     }
 
     /**
+     * Waits until UI has settled down
+     */
+    void waitUntilReady() {
+        TitledPane titledPane = getPane();
+        waitForStatus(titledPane, READY_STATUS);
+    }
+
+    /**
+     * Retrieves the title pane
+     *
+     * @return the selected pane
+     */
+    private TitledPane getPane() {
+        return getElementById(HISTORY_PARENT, TitledPane.class);
+    }
+
+    /**
+     * Get state from titlepane
+     */
+    boolean titlePaneShows(String msg) {
+        TitledPane titledPane = getPane();
+        return titledPane.getText().equals(
+                I18n.getMsg(msg));
+    }
+
+    /**
+     * Sets the serviceURL
+     */
+    void setServiceUrl(String url) {
+        TextField serviceURL = getServiceURL();
+        serviceURL.setText(url);
+    }
+
+    /**
+     * Retrieve ServiceURLTextfield
+     *
+     * @return Textfield
+     */
+    private TextField getServiceURL() {
+        return getElementById(SERVICE_URL, TextField.class);
+    }
+
+
+    /**
      * Gets Element from the scene.
+     *
      * @param element Selector for the element to look
-     * @param type Class of the Element
-     * @param <T> Generic returntype
+     * @param type    Class of the Element
+     * @param <T>     Generic returntype
      * @return Returns Element
      */
-    <T extends javafx.scene.control.Control> T getElementByName(
+    private <T extends javafx.scene.control.Control> T getElementById(
             String element, Class<T> type) {
         return type.cast(scene.lookup(element));
     }
